Refactor some code to remove duplication
