+++ foreshadow/foreshadow.py
@@ -17,10 +17,10 @@ from foreshadow.serializers import (
     _make_deserializable,
     _make_serializable,
 )
-from foreshadow.utils import check_df, get_transformer
+from foreshadow.utils import CustomizeParamsMixin, check_df, get_transformer
 
 
-class Foreshadow(BaseEstimator, ConcreteSerializerMixin):
+class Foreshadow(BaseEstimator, ConcreteSerializerMixin, CustomizeParamsMixin):
     """An end-to-end pipeline to preprocess and tune a machine learning model.
 
     Example:
@@ -328,7 +328,7 @@ class Foreshadow(BaseEstimator, ConcreteSerializerMixin):
 
         """
         params = self.get_params(deep)
-        selected_params = self.__create_selected_params(params)
+        selected_params = self.customize_serialization_params(params)
         serialized = _make_serializable(
             selected_params, serialize_args=self.serialize_params
         )
@@ -406,7 +406,7 @@ class Foreshadow(BaseEstimator, ConcreteSerializerMixin):
         estimator.set_params(**data)
         return estimator
 
-    def __create_selected_params(self, params):
+    def customize_serialization_params(self, params):
         """Extract params in the init method signature plus the steps.
 
         Args:
@@ -416,12 +416,7 @@ class Foreshadow(BaseEstimator, ConcreteSerializerMixin):
             dict: selected params
 
         """
-        init_params = inspect.signature(self.__init__).parameters
-        selected_params = {
-            name: params.pop(name)
-            for name in init_params
-            if name not in ["self", "kwargs", ""]
-        }
+        selected_params = super().customize_serialization_params(params)
         selected_params["data_columns"] = params.pop("data_columns")
         return selected_params
