+++ foreshadow/utils/__init__.py
@@ -2,6 +2,8 @@
 
 
 from foreshadow.utils.common import (
+    ConfigureColumnSharerMixin,
+    CustomizeParamsMixin,
     get_cache_path,
     get_config_path,
     get_transformer,
@@ -30,4 +32,6 @@ __all__ = [
     "is_transformer",
     "is_wrapped",
     "dynamic_import",
+    "ConfigureColumnSharerMixin",
+    "CustomizeParamsMixin",
 ]
